Change arguments handling
Allow `-` to name stdin or stdout.
Input file is always the first argument, no need to use option `-i`.
Output file can be named with `-o` or as the second argument
but not both. Option `-o` is useful when there is no input file (input
is redirected, for example: `mysql-to-sql.py -o out.sql < in.sql`).